Split to different files issue #39
